Added error package and business logic

+++ pkg/model/collection.go
@@ -2,6 +2,6 @@ package model
 
 type Collection struct {
 	Items      interface{} `json:"items"`
-	PageNumber int64       `json:"page_number"`
-	PageSize   int64       `json:"page_size"`
+	PageNumber int         `json:"page_number"`
+	PageSize   int         `json:"page_size"`
 }
